refactor(parquet-datasource): split opener.rs into an opener/ module
Pure code motion, no behavior change and no public API change
(`opener` was already a private module). Splits the ~2,700 LOC
`opener.rs` into a directory module:

- `opener/early_stop.rs` — `EarlyStoppingStream`, the dynamic-filter
  early-termination wrapper applied at the end of `build_stream`.
- `opener/encryption.rs` — `EncryptionContext` and the
  `ParquetMorselizer::get_encryption_context` helpers, isolating the
  `#[cfg(feature = "parquet_encryption")]` gating that previously bled
  through the main file.

`opener.rs` becomes `opener/mod.rs`.

Split out of #22156, which originally also extracted an
`opener/push_decoder_stream.rs`; that move is now obsolete since #22289
already extracted `PushDecoderStreamState` into `push_decoder.rs`.
